True fix of parameterize issue
It turned out the fix in https://github.com/shirok/WiLiKi/commit/f7d9494d31f9af88889627b14b25413725fb628e was missing the target.
The true cause was that we switched the semantics of 'delay' in 0.9.13;
before, it was R7RS semantics where the promise is evaluated in the
dynamic environment of the first 'force', while, in 0.9.13 we adopt
SRFI-226 semantics where the promsie is evaluated in the dynamic
environment of 'delay'.  Since the database is not open in the
dynamic environment of 'delay', we got 'database is not open' error.